spawning 10 in paralel works
